more help for autumn
now, whenever the diff in daylight is negative for either morning or
evening, we're showing the next date when the diff will be the same.

This will help me during runs in the morning to tell me "oh - just wait
till March and it'll be as dark as it is now".